power: bq27541: Eliminate qpnp-fg dependency
Since we have our own dedicated fuel-gauge hardware and driver (bq27541),
the qpnp-fg driver doesn't provide any extra features or data that is
useful to us. Furthermore, the qpnp-fg hardware fires a lot of interrupts
for "mem-avail" and holds wakelocks to perform tasks that provide nothing
for us.

Our bq27541 fuel-gauge IC is providing all of the core battery monitoring
functionality, so remove the qpnp-fg dependency since it's just a resource
hog for us. This change makes the bq27541 driver emulate the qpnp-fg driver
by using "bms" for the power supply name in order to maintain compatibility
with existing drivers.

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com> (reverted from commit 59d4d89a48a3b1cfdd110a66f08cfaabc1005584)